Use Vert.x as the basis for all HTTP
This moves to the new Vert.x based Undertow branch. All HTTP trafic
is not routed through Vert.x, with Undertow providing a Servlet
implementation on top of this.

This allows for a common HTTP layer where all HTTP endpoints can
be configured in a unified manner, including things like security,
logging etc.